+++ .ts-for-gir.gcalc.rc.js
@@ -1,8 +1,7 @@
 export default {
+    environments: ['gjs', 'node'],
     modules: [/*'GCalc-1',*/ 'GCalc-2'],
     girDirectories: ['./vala-girs/gir-1.0'],
-    pretty: false,
-    noCheck: false,
     ignore: [],
     ignoreVersionConflicts: true
 }