+++ .ts-for-gir.gtk3.rc.js
@@ -1,7 +1,6 @@
 export default {
+    environments: ['gjs', 'node'],
     modules: ['Gtk-3.0'],
     girDirectories: ['./vala-girs/gir-1.0'],
-    pretty: false,
-    noCheck: false,
     ignore: [],
 }