+++ .ts-for-gir.modemmanager1.rc.js
@@ -1,7 +1,6 @@
 export default {
+    environments: ['gjs', 'node'],
     modules: ['ModemManager-1.0'],
     girDirectories: ['./vala-girs/gir-1.0'],
-    pretty: false,
-    noCheck: false,
     ignore: [],
 }