+++ .ts-for-gir.rygelcore2.rc.js
@@ -1,7 +1,6 @@
 export default {
+    environments: ['gjs', 'node'],
     modules: ['RygelCore-2.6', 'RygelRenderer-2.6'],
     girDirectories: ['./vala-girs/gir-1.0'],
-    pretty: false,
-    noCheck: false,
     ignore: [],
 }