+++ packages/cli/templates/Gjs/ui/calendar.js
@@ -1,0 +1,5 @@
+<% if(moduleType === 'esm') { %>
+    export imports.ui.calendar;
+<% } else { %>  
+    module.exports = imports.ui.calendar;
+<% } %>