+++ packages/cli/templates/Gjs/ui/messageList.js
@@ -1,0 +1,6 @@
+<% if(moduleType === 'esm') { %>
+    export imports.ui.messageList;
+<% } else { %>  
+    module.exports = imports.ui.messageList;
+<% } %>
+