+++ NEWS.md
@@ -5,14 +5,16 @@
 - Define required .gir files for GNOME Shell type generation (e.g. Clutter-11.gir / Clutter-10.gir, St-1.0.gir, Shell-0.1.gir, ...)
 - Add support to generate packages for the types we can publish on NPM 
 
-# Unreleased
+# 3.0.0-beta.9
+- Add support for the `"gi://..."` import syntax, see #103 by @CharlieQLe
 - README.md greatly simplified and split into several files
 - Add new DependencyManager to be able to search for a dependency regardless of the version 
 - Improved GNOME Shell types
 - Only extend class/interface parent if the dependency of the parent exists
-- Add more properties to the dependency object like `exists` to be able to check if a dependency gir file was found
+- Add more properties to the dependency object like `exists` to be able to check if a dependency `gir` file was found
 - Upgrade dependencies
 ## Breaking Changes
+- Removed the `pretty` and ` noCheck` options
 - Changed default options to:
   - `noNamespace: true` (was `useNamespace: false` before)
   - `environments: ["gjs"]` (was `["gjs", "node"]` before)